Moved run_if_baspacho to theseus.extlib.tests.common.py.
