+++ theseus/optimizer/linear/tests/test_baspacho_sparse_solver.py
@@ -9,7 +9,7 @@ import numpy as np
 
 import theseus as th
 
-from theseus.constants import run_if_baspacho
+from theseus.extlib.tests.common import run_if_baspacho
 from theseus.utils import random_sparse_binary_matrix, split_into_param_sizes
